refactor(pinyin): split pinyin.ts (593 LOC) into focused modules (Related to #1957)
Extract static data and algorithms from monolithic pinyin.ts into:
- pinyin/data.ts — PINYIN_GROUPS, charToPinyin, NEAR_SOUND_PAIRS, STT_EQUIVALENT_GROUPS
- pinyin/normalize.ts — getPinyin() character-to-toneless-pinyin lookup
- pinyin/match.ts — isHomophone, isNearSound, isSttEquivalent, correctHomophones
- pinyin.ts — thin re-export facade (zero logic, backward compat)

TDD: 38 vitest tests written before refactor, all pass after.
No changes to public API; all 3 consumers (textDiff, fluencyAnalyzer,
liveTutorHelpers) unmodified.
